requirements.txt
open_clip_torch 
matplotlib
numpy
flask
unidecode
pillow
googletrans==4.0.0-rc1
